feat(web): new CYA back link approach
We now save submitted values to a special session key (yourKey/edit). This
allows us to roll back the edits if the user clicks back through the flow
and ends up on the CYA page without completing the edit. This is necessary
because if you have changed the values for one page (e.g. changed no to yes)
then the fields on the next page may be required (e.g. address details),
but if you click back after editing only some pages then the form could
contain invalid data, so we need to reset it to that previously valid values.

A2-3670