fix(api): improve API performance for GET /appeals endpoint
Use DB-level paging to improve performance. We were getting all records and then splitting
them up in the code. The query that it was generating actually broke the DB when there were
more than 2100 records. We now only get one page of records and do a separate COUNT query
to get the total number of matched records for paging purposes. Also I have simplified the
query for getting all unique statuses fron the national list as that was also broken when
there were > 2100 records.

A2-3901