fix(api): make sure all values for filter dropdowns are returned
The `GET /appeals` endpoint returns some extra values for populating the
filter dropdowns. Now that we are paging the results at the DB level, only
the values for appeals on the current page were showing in some of the
dropdowns. So we do need to grab all of the appeals for the purpose of
knowing which values to show, but that can be done as a second query without
including all of the relationships for all of the appeals, which makes it
more performant and stops it breaking when there are many appeals in the DB.
A couple of extra queries are required to translate IDs/keys into names but
there are not too demanding. At some point in the future we will be looking
at using stored procedures as an option for further improving the performance
here.

A2-3901