fix(api): roll back status transitions properly
We need to keep a linear history of appeal status transitions. There are some
situations where an appeal might have to transition back to a state that it
has had previously. In this case we need to find that old status, set it to valid,
and delete any ones that were created after that.

A2-4126