feat(api): add OpenAPI docs check to pre-commit hook
I't svery easy for people to forget to run `npm run gen-api-spec` before
committing their changes. This results in a drift between the intended spec
and the generated files. The next person who comes along to run the script
will end up with not only their cahnges, but the changes other people have
made since the script was last run. Even worse, sometimes people edit the
generated files directly, and their changes get overwritten when the script
is run later on.

To prevent these issues from happening I have introduced a script which runs
in a pre-commit hook which runs the `gen-api-spec` script, but outputs the
generated files to a temporary directory, and diffs the files with the
existing ones. If they do not match it shows the diff along with an error
message and prevents the code from being committed.